feat(storage): merge notebooks and datasets in data source panel
