+++ packages/rath-client/src/pages/dataSource/selection/index.tsx
@@ -13,7 +13,7 @@ import OLAPData from './olap';
 import HistoryPanel from './history';
 import DatabaseData from './database/';
 import AirTableSource from './airtable';
-import Notebook from './notebook';
+import KanariesCloud from './cloud';
 
 interface SelectionProps {
     show: boolean;
@@ -33,8 +33,8 @@ const Selection: React.FC<SelectionProps> = props => {
     const dsTypeOptions = useDataSourceTypeOptions();
 
     const formMap: Record<IDataSourceType, JSX.Element> = {
-        [IDataSourceType.NOTEBOOK]: (
-            <Notebook setLoadingAnimation={setLoadingAnimation} />
+        [IDataSourceType.CLOUD]: (
+            <KanariesCloud setLoadingAnimation={setLoadingAnimation} />
         ),
         [IDataSourceType.FILE]: (
             <FileData onDataLoading={onDataLoading} onClose={onClose} onDataLoaded={onDataLoaded} onLoadingFailed={onLoadingFailed} toggleLoadingAnimation={setLoadingAnimation} />
