+++ packages/rath-client/src/store/dataSourceStore.ts
@@ -3,9 +3,8 @@ import { IReactionDisposer, makeAutoObservable, observable, reaction, runInActio
 import { combineLatest, from, Observable, Subscription } from "rxjs";
 import { getFreqRange } from "@kanaries/loa";
 import * as op from 'rxjs/operators'
-import { BlobWriter, TextReader, TextWriter, ZipReader, ZipWriter } from "@zip.js/zip.js";
 import { notify } from "../components/error";
-import { KanariesDatasetFilenameCloud, KanariesDatasetPackCloud, RATH_INDEX_COLUMN_KEY } from "../constants";
+import { RATH_INDEX_COLUMN_KEY } from "../constants";
 import {
     IDataPreviewMode,
     IDatasetBase,
@@ -25,9 +24,6 @@ import {
     IBackUpData,
     ICreateDataSourcePayload,
     ICreateDataSourceResult,
-    IDatasetMeta,
-    IDatasetData,
-    ICreateDatasetPayload,
 } from "../interfaces";
 import { cleanDataService, filterDataService,  inferMetaService, computeFieldMetaService } from "../services/index";
 import { expandDateTimeService } from "../dev/services";
@@ -1165,65 +1161,6 @@ export class DataSourceStore {
         } catch (error) {
             console.error(error);
             return null;
-        }
-    }
-
-    public async saveDatasetOnCloud(payload: ICreateDatasetPayload): Promise<{ downloadUrl: string } | null> {
-        const createDatasetApiUrl = getMainServiceAddress('/api/ce/dataset');
-        const reportUploadSuccessApiUrl = getMainServiceAddress('/api/ce/upload/callback');
-        const zipFileWriter = new BlobWriter();
-        const zipWriter = new ZipWriter(zipFileWriter);
-        const data: IDatasetData = {
-            data: await this.backupDataStore(),
-            meta: await this.backupMetaStore(),
-        };
-        const content = new TextReader(JSON.stringify(data));
-        zipWriter.add(KanariesDatasetFilenameCloud, content);
-        const blob = await zipWriter.close();
-        const file = new File([blob], KanariesDatasetPackCloud);
-        try {
-            const createDatasetApiRes = await request.post<ICreateDatasetPayload, { storageId: number; uploadUrl: string }>(
-                createDatasetApiUrl, payload
-            );
-            const fileUploadRes = await fetch(createDatasetApiRes.uploadUrl, {
-                method: 'PUT',
-                body: file,
-            });
-            if (!fileUploadRes.ok) {
-                throw new Error(`Failed to upload file: ${fileUploadRes.statusText}`);
-            }
-            const reportUploadSuccessApiRes = await request.get<{ storageId: number; status: 1 }, { downloadUrl: string }>(
-                reportUploadSuccessApiUrl, { storageId: createDatasetApiRes.storageId, status: 1 }
-            );
-            return reportUploadSuccessApiRes;
-        } catch (error) {
-            console.error(error);
-            return null;
-        }
-    }
-
-    public async loadDatasetOnCloud(dataset: IDatasetMeta): Promise<boolean> {
-        const { downloadUrl } = dataset;
-        try {
-            const data = await fetch(downloadUrl, { method: 'GET' });
-            if (!data.ok) {
-                throw new Error(data.statusText);
-            }
-            if (!data.body) {
-                throw new Error('Request got empty body');
-            }
-            const zipReader = new ZipReader(data.body);
-            const file = (await zipReader.getEntries()).find(entry => entry.filename === KanariesDatasetFilenameCloud);
-            if (!file) {
-                throw new Error('Dataset file not found');
-            }
-            const writer = new TextWriter();
-            const dataset = JSON.parse(await file.getData(writer)) as IDatasetData;
-            await this.loadBackupDataStore(dataset.data, dataset.meta);
-            return true;
-        } catch (error) {
-            console.error(error);
-            return false;
         }
     }
